refactor: abstract ctrlk to a component
