+++ app/[locale]/header.tsx
@@ -2,9 +2,10 @@ import { headers } from 'next/headers';
 import Image from 'next/image';
 import Link from 'next/link';
 import { userAgent } from 'next/server';
+import { FaMagnifyingGlass } from 'react-icons/fa6';
 import { IoMenu } from 'react-icons/io5';
 
-import SearchButton from '~/app/search';
+import { CtrlLink } from '~/components/link';
 import LocaleSwitcher from '~/components/locale-switcher';
 import { Button } from '~/components/ui';
 import { getTranslations } from '~/i18n/translations';
@@ -103,11 +104,30 @@ export default async function Header({ locale }: { locale: string }) {
             </Button>
           </li>
           <li>
-            <SearchButton
-              href={`/${locale}/search`}
-              text={text.search}
-              isMacOS={isMacOS}
-            />
+            <Button asChild className="xl:hidden" variant="icon">
+              <CtrlLink href={`/${locale}/search`} shortcut="k">
+                <FaMagnifyingGlass className="p-2 text-primary-700" size={40} />
+              </CtrlLink>
+            </Button>
+
+            <Button
+              asChild
+              className={cn(
+                'hidden xl:flex',
+                'group h-full w-60 gap-3 bg-neutral-50 px-4'
+              )}
+              variant="secondary"
+            >
+              <CtrlLink href={`/${locale}/search`} shortcut="k">
+                <FaMagnifyingGlass size={16} />
+                <span className="grow text-left text-neutral-500 group-hover:text-neutral-100">
+                  {text.search}
+                </span>
+                <kbd className="font-sans font-medium opacity-50">
+                  {isMacOS ? '⌘' : 'Ctrl'} K
+                </kbd>
+              </CtrlLink>
+            </Button>
           </li>
           <li className="hidden lg:block">
             <Button asChild className="h-full w-16 xl:w-20">
